+++ src/objects/Datacenter.js
@@ -1,11 +1,13 @@
 import z from 'zod';
 import Guard from '../utils/guard.util.js';
+import Agent from './Agent.js';
 
 export class Datacenter {
   #id_datacenter;
   #label;
   #provider;
   #city;
+  #agent;
 
   constructor(props) {
     const data = Guard.validateProps(Datacenter.schema, props);
@@ -13,6 +15,7 @@ export class Datacenter {
     this.#label = data.label;
     this.#provider = data.provider;
     this.#city = data.city;
+    this.#agent = data.agent;
   }
 
   // Zod Schema for object validation
@@ -21,6 +24,7 @@ export class Datacenter {
     label: z.string().min(1).trim().default(''),
     provider: z.string().min(1).trim().default(''),
     city: z.string().min(1).trim().default(''),
+    agent: z.instanceof(Agent).default(new Agent({})),
   });
 
   // Getters
@@ -40,6 +44,10 @@ export class Datacenter {
     return this.#city;
   }
 
+  get agent() {
+    return this.#agent;
+  }
+
   // Setters
   set id_datacenter(value) {
     this.#id_datacenter = value;
@@ -56,6 +64,9 @@ export class Datacenter {
   set city(value) {
     this.#city = value;
   }
+  set agent(value) {
+    this.#agent = value;
+  }
 
   public_format() {
     const format = {
@@ -63,6 +74,7 @@ export class Datacenter {
       label: this.#label,
       provider: this.#provider,
       city: this.#city,
+      agent: this.#agent.toJSON(),
     };
     return format;
   }
@@ -73,6 +85,7 @@ export class Datacenter {
       label: this.#label,
       provider: this.#provider,
       city: this.#city,
+      agent: this.#agent.toJSON(),
     };
   }
 }